Refactor JWT handling by moving related functions and classes to a new jwtoken module, removing deprecated admin routes and cleanup scripts.
