+++ routes/picture.py
@@ -3,7 +3,7 @@ import os
 from flask import Blueprint, jsonify, request, send_from_directory
 
 from config.settings import Config
-from jwt_helper import token_required
+from jwtoken.decorators import token_required
 from utility.database import database_cursor
 
 picture_blueprint = Blueprint("picture", __name__)
