+++ tests/test_jwt/conftest.py
@@ -2,7 +2,7 @@ import jwt
 import pytest
 from flask import Flask
 
-from jwt_helper import generate_access_token
+from jwtoken.tokens import generate_access_token
 
 
 @pytest.fixture
