+++ tests/test_routes/test_authentication/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 from flask import Flask
 
-from jwt_helper import generate_access_token
+from jwtoken.tokens import generate_access_token
 from routes.authentication import authentication_blueprint
